refactor: move webapi utility modules
