+++ src/image_annotator_lib/webapi/__init__.py
@@ -1,1 +1,2 @@
-"""WebAPI-specific implementation modules."""
+"""WebAPI execution support package."""
+
