+++ src/image_annotator_lib/webapi/http_retry.py
