+++ src/image_annotator_lib/webapi/model_id.py
