+++ tests/unit/core/test_http_retry_transport.py
@@ -17,7 +17,7 @@ import pytest
 from pydantic_ai.retries import AsyncTenacityTransport
 from tenacity import wait_none
 
-from image_annotator_lib.core.http_retry import (
+from image_annotator_lib.webapi.http_retry import (
     HTTP_RETRY_MAX_ATTEMPTS,
     HTTP_RETRY_MAX_WAIT_SECONDS,
     build_retry_transport,
